Sửa giấy một chút
Sửa bảng giá giấy

+++ TinhGiaInBDO/CanGapBDO.cs
@@ -10,6 +10,6 @@ namespace TinhGiaInBDO
     {
 
         public int TocDoConGio { get; set; }
-        public int MotDuongCanTangThoiGianChuanBi { get; set; } //phần so với đường kế cận
+        public float MotDuongCanTangThoiGianChuanBi { get; set; } //phần so với đường kế cận
     }
 }
